+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 4.2.1 / Unreleased
 - Install pam_dc dependencies using extra syntax: `pip install wsgidav[pam]`
 - #281 Requesting range off end of file does not return 416 status code
+- #282 Hotfix PUT request without content-length (fix for Finder on MacOS Ventura)
 
 ## 4.2.0 / 2023-02-18